feat: enhance logging functionality with trace controls and diagnostics integration #262
